Refactor pseudo-blitting to use only integers.
Previously we have used an Array<String> representing a single image, and that
string was converted into an array of integers. Obviously this isn't very
efficient and as we naw have a sprite/animation file reader, we no longer need
to do any conversion at runtime because we can actually "bake" in these values.

This also means, that the draw_string() and draw_block() methods don't exist
anymore, as they obviously didn't make sense so far, aka "reserved for future
use". But even in the future, we still can use blit() for these operations.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>