Object: Automatically grow parent objects.
This only is done if autoresize is true in order to prevent accidentally growing
the root canvas. This should fix a few issues with sprites being cut off when
new children are added.

Although there still is another bug which cuts off the width of _all_ sprites,
but let's look forward to the fix in a future commit.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>