Add test system with simple test cases.
If we are going to refactor the sprite engine we may want to make sure that
everything draws correctly. Fortunately for us, ASCII "graphics" are really easy
to test, so our test system is based on a test canvas (a simple array of
character codes) and compares the result of the sprite manipulation from test
canvas with the expected value as an array of strings.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>