Initial support for dirty rectangle based sprites.
Although there still is some flickering, we pretty much fix the failing test
cases and prevent artifacts to be left behind after a sprite redraw.

The rect support currently is somewhat bloated, as it contains methods for
future use[TM], so we might want to remove them if we don't need it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>