Fix dirty flag to be unset only in blit().
We _always_ want to run the update() methods, because they're actually the
reason _anything_ happens. The only thing we want is to avoid redrawing too
much.

And of course we want to propagate the dirty flag to all children, because,
let's say the x value has changed, then we need to redraw the children as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>