Register sprites as an array in root Surface.
If we want to do dirty rectangle drawing properly, we need to have all sprites
and their corresponding rects in one place, even though the sprites have a
different nesting.

That way we populate Surface.sprites as soon as a new Sprite is attached and
remove it once it's detached. Of course we need to propagate through all
children, because Sprite objects can be attached to themselves without a proper
surface.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>