Surface: Refactor to use dirty rectangle drawing.
The _real_ blit function is now in Surface and the Sprite only triggers a
refresh of the current rectangle combined with the previous rectangle and passes
the union to the Surface object.

Within the Surface object, we check for all sprites that intersect with this
union rectangle and trigger an update of all the intersecting sprites *and* the
sprite within the union (because it intersects as well) in the correct order of
their corresponding Z value.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>