Animation: Factor out asset reader into macro.
First of all, the from_file() static method is now at the bottom of the
Animation class in order to keep important stuff at the top of the file.

In addition to this, Animation now gets passed an additional parameter which
includes an array of options for each frame. This makes it possible to add
special annotations to our asset files.

For now there is only "reference", but this is unparsed by now and will soon be
a character that is recognized as the reference point for positioning the
sprite.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>