macro.Animation: Correctly apply/parse refpoints.
Finally we don't have these damn "x"es from the wing data files in the
rendering. Reference points should mark a point in the image with an unique
character in order to center the sprite around that point. This should make wing
animation possible very soon.

The only thing that's missing is the runtime part, to correctly draw the sprite
centered on the reference point.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>