Replace CP437 font with an inlined bitmap version.
The new way of rendering fonts is by using an array of bit shifted 16 bit
integers and every character is written into an offscreen canvas
(TermCanvas.font_canvas) and blitted off that canvas.

In order to support colors we need this kind of mechanism, because the cp437.png
file contained 8bit data rather than a bitmap (not a big problem though), but we
need to have the bitmap separated by foreground and background masks (remember,
in VTs you can have foreground _and_ background colors) and this operation will
cost us, except if we render it to an offscreen canvas as well.

And doing it this way has two other advantages:

 * We can easily modify the font.
 * It is just one JS and a HTML, nothing more to ship.

For the point of supporting colors: If you look at the add_to_font_cache()
method, you will see that we have 0xBEBEBE there, so it's really a breeze to add
colors. Especially because we just need to multiply the color value with the
ordinal and add it to the hashtable.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>