Migrate to Haxe 3 release canidate.
As pointed out in the color support merge commit (b145b77), it really makes
sense if we fully switch to Haxe 3. Primarily because it makes a _bunch_ of
things way easier for us and on the other hand, I guess our Open Source release
won't happen before the final version of Haxe 3 is released anyway.

This change is actually just doing the basic things pointed out at:

http://haxe.org/manual/haxe3/migration

No rewrite of the macros or any abstract type implementation is done right now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>