Add a Matrix abstract type for 2D array handling.
This should get rid of most of the repetitive width and height calculation
functionality and also simplifies the underlying data structure quite a lot, as
the matrix is actually a one dimensional array holding all the values.

But on the contrary, this will make our macro functionality much more complex
and generates a lot of unnecessary code in the output, so we need to refactor
our animation file parsing code. Fortunately, this is on our roadmap :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>