Matrix: Allow for growing/shrinking and add tests.
This should at least make our matrix implementation useful for populating data,
but we probable want to have a mutable and a immutable version? But for now, we
only need a mutable version and the implementation has to be well tested.

In addition, we now have a stub inline function get_default_value(), which
assigns the right default value for the corresponding type, so we no longer
should have null values (except for types without a sensible default) whenever
we need to fill rows or columns.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>