Simplify AnimationParser down to the essentials.
In order to rewrite the animation file parser, we first need to simplify a few
things:

 * Get rid of all those nested Expr by using macro reification in Haxe 3.
 * Don't support arbitrary options and only handle reference points by now.
 * Simplify options structure of Animation to a single array of objects.

Especially by not supporting arbitrary options, this gets rid of quite a lot of
cruft, like macro.Types. As our goal for the new animation file format are
generic containers anyway, we don't need to have such a complexity in the first
place.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>