AnimationParser: Use a simpler container format.
Prefixing each frame with its number was never needed and just left quite a lot
of possibilities where bugs could happen (for example if the number doesn't
start with 0 or there are holes in between the numbers). This is now obsolete
and every frame just starts with a vertical bar.

But that's not the only thing that's changed, so we now have a separation
character in-between frames, because every frame can contain multiple container
for other types of images, that correspond to the frame. One particular example
are color mappings, where we already have a variant though it's unused.

With this refactoring we now also have some simple test cases. While these are
by no means exhaustive, they at least make it easier to prototype and test the
final animation file format.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>