AnimationParser: Add new separate FrameAreaParser.
This is mainly to separate the parsing of "drawing area" inside the animation
files, as this is the part where we want to have containers, as in boxes
surrounded by horizontal and vertical characters, like this:
  _____
 |     |
 |     |
 |_____|

As some people might notice in this code, I originally wanted to go for a much
higher aim, having some generic container format. But in reality, we don't need
to have this, as we're going to hell if we want to support boxes within boxes.

So, for now, let's head on to the simple format, having two types of containers:

: heading :
| content |

and the second type is just:

| content

notice the missing vertical bar at the end? this is by intention, because it
makes simple frames as they should be: simple.

And because we're still not finished with our parser implementation, some test
cases do not succeed at the moment.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>