FrameAreaParser: Refactor/finish implementation.
This one underwent a major design decision to use : for header fields, because
using just | would be way too ambiguous, especially if the data in the drawing
area of a container looks like the container itself.

The method we use here is to detect horizontal borders (consisting of | and :)
in the whole drawing area, which then are grouped together based on whether
they're adjacent to each other and match in size.

After we have a collection of grouped borders, we merge every border into Rect
objects of their inner areas and make sure we throw away the ones that are
overlapped by parent Rect objects that have a lower Y value.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>