Symbol: Properly handle default values for colors.
This a somewhat pesky implementation, because we're converting from a plain
integer value to a Symbol. Basically, this means, that all the color bits are
set to 0 for integers casted to a Symbol.

Unfortunately the best way to solve this is by getting rid of the cast in
geom.Matrix, which isn't possible without quite a lot of duplicated code. So the
best way to fix this is within Symbol, by adding two additional bit fields for
"fg color set" and "bg color set".

Another approach in solving this would be to swap 7 by zero in the xterm color
table. Though this might work as well, it isn't a feasible implementation as it
might cause confusion if you expect an xterm color palette while it's not due to
this swap.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>