New Color class for converting RGB/xterm colors.
At the moment, this only consists of a static method to convert a xterm color
into an RGB value. As fonts are cached within the (extra) canvas, we don't want
to use a macro to bloat the output code. That's why the testing is done by
simply matching with a static color table from the XTerm source code:

http://invisible-island.net/xterm/

Signed-off-by: aszlig <aszlig@redmoonstudios.org>