AnimationParser: Correctly merge all containers.
Currently this only supports 6x6x6 cube values of the 256 color XTerm palette,
but for now this is sufficient enough. If we really need to have shades of gray,
we can still add a new container type.

Merging of containers is done mutually exclusive, which means: If you already
have a 256 color value set, you can't have an ANSI color for the particular
character and vice versa.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>