Timer: Refactor and skip frames more efficiently.
In order to simplify things, we now have one get_frame_handler() method that
deals with the cross-browser fuckery and everything else is not touching this
part and utilizes the type system. So the only thing that could go wrong in
respect to types is within this method.

Another thing which this refactoring brings is to avoid rendering frames if the
delta is bigger than ten times the frame time.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>