Rect: Refactor as an abstract type based on array.
We now represent rectangles based on a 4 element array like this:

  [x, y, width, height]

That way we don't need the toString() method anymore as the resulting string
representation always is in the right order.

And of course this should speed up rectangle handling significantly if there are
a lot of sprite objects in the scene/surface.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>