Rename everything from "jascii" to "hase".
The acronym "hase" stands for "Haxe Ascii Spriting Engine".

This means that jascii will still exist, but as a separate project which is
using hase for display. Hase itself will only be the spriting and animation
engine and that's why the splitup.

Jascii as a mix between Jabber (XMPP) and ASCII obviously won't fit for the
spriting engine itself. And even jascii is just a working title at the moment.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>