Add a license (MIT/Expat) to the project.
This should ensure a maximum of compatibility with other licenses including the
GPL. My original intent differed from this, as i wanted to prevent corporations
from using the library without publicly stating that they're doing so. But in
the end I don't think that this piece of software is substantial enough to be
worthwhile for a big company to build on it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>