AppBuilder: Don't overwrite unrelated styles.
Previously we modified the padding and margin of the body element, now we are
just using "position: fixed" and set values _only_ on the newly created canvas
element.

Now it's more easy to provide content in the page's body in case we're runing on
browsers either not supporting the canvas element or not supporting JavaScript
altogether.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>