appbuilder: Generate target specific main() code.
This involves a bit of ugliness in the build function, to ensure we generate the
right fields for the specific target as we don't have the target constant
available during macro evaluation.

Of course, this integrates our newly created TermCurses class, so the cpp should
give us a somewhat useful output.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>