object: Fix and test Z ordering.
We're going to add some kind of pseudo double buffering to the surface blitter
soon[TM], so we need to make sure that Z ordering works the way we want it to.

It did so far, but only when the Z value is set _after_ adding the object onto
another one. The test case now added makes sure that either this issue nor some
related issue (that might come up soon) won't happen again.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>