Move everything terminal related into hase.term.
This is to allow for seamless abstraction of the terminal and to make it easy to
share code between the different terminal types without cluttering the hase
namespace. And of course, the naming is now more consistently and we call
everything, be it a "surface provider" or a "term canvas", simply "terminal" no
matter if it's drawn using pixel operations, by sending ECMA-48 CSIs or just a
test terminal.

Everything in hase.term.internal is now only accessible from the hase.term
namespace, so that the name is clearly appropriate.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>