surface: Refactor redraw_rect() and blitting.
We now no longer redraw everything in the universe and only the relevant parts
within the dirty range, like this:

  0123456789012345678
  1                    A (5,  2)
  2     ____           : (8,  4)
  3    |    |          B (12, 6)
  4    |  ..|....
  5    A__:c|   :      : x A (8,  4, 2, 1)
  6       :    _:__    : x B (12, 6, 2, 1)
  7       :...|c:  |
  8           |    |   c   -> common area
  9           |____B   :   -> rectangle to redraw
  0                    A/B -> sprites

This basically shows the dirty range in dots (the rectangle to redraw) and the
intersecting sprites A and B.

Our current implementation iterates through all registered sprites, throwing
away non-visible sprites and those out of range. Sprites that are in range are
then applied to a base matrix, ignoring alpha characters. Order of appliance is
based on the Z value of the corresponding sprites and we're drawing from bottom
to top.

Of course, the way of drawing is not very optimal, especially if there are a
whole lot of overlapping sprites. But for now this solution is the most accurate
and that's what we're testing for, we can still optimize later.

Speaking of testing, this commit fixes test_different_overlapping_frame_sizes()
as well, as - shame on me - I copied the result directly into the test output,
assuming it was correct.

I've noticed this bug, because sometimes there were drawing artifacts visible,
which only happened in the example when the framerate dropped.

Background of the bug is as follows:
  ___      ___        ___
 |   |    |   |      |   |
 | A |    | A |_     | A_|_
 |___|    |___| |    |_|   |
  ___  ->  _|_C | ->  _|_C |
 |   |    |   |_|    |   |_|
 | B |    | B |      | B |
 |___|    |___|      |___|

What happens is the following:

 1. A and B are in the scene, this is the starting point.
 2. C is added with a lower Z value and thus A and B get redrawn on top of it
    because they're *both* intersecting with C.
 3. Ugly things happen now as soon as we either update A or B, because they only
    intersect with C and won't trigger a redraw of the respective counterpart.

As explained in the beginning, this shouldn't happen anymore, as we only redraw
the characters intersecting with the dirty range, so A and B only redraw the
corner of C that overlaps with them, nothing more.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>