surface: Clean up and optimize redraw_rect().
In the previous commit, we just tried to get the new rectangle redrawer running
and pass all the test. With this commit, it actually is now way faster (around
20% CPU usage on my system with curses, in comparison to 80% before) because we
don't iterate through the whole matrix, even if we only want to set a small
fraction on the base matrix.

And of course, this should also clean up the mess/ugly code we had until now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>