surface: Prevent and test area from overflowing.
This actually has been an off-by-one error which led to ugly but barely
notable artifacts when used with the curses terminal. For JS we set a
specific canvas size, so overflowing the canvas will simply lead to the
content not being shown.

We also prevent this from happening again in the future by adding two
test cases for each direction. As the test terminal is also just drawing
stuff in an array, we just resize the surface to stay within the bounds
of the test terminal, as the behaviour is quite undefined there as well
as in real-world implementations.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>