matrix: Refactor from abstract to a class.
After messing around in 12(!) different branches creating default values
for that abstract type, getting to realize that I was switching to
creating a class anyway using macros.

So, why not make it a class in the first place and use inline functions
to operate on it? There is really no reason to argue against that,
because it's simpler, more explicit but also has advantages in terms of
performance.

With the abstract type, we used an anonymous structure as its base type,
which performs dynamic lookup on static platforms.

This removes fold_maxsize(), which was unused since c638d0d and
from_internal(), which was only used internally by the abstract type.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>