Animation: Fix Matrix macro serialization.
We can't push the parsed animation frame data using our new Matrix type,
because it's not an anonymous structure anymore.

It's a bit of a setback here, because we have to convert back and forth
to/from Dynamic, but in the end we can only do platform specific stuff,
such as initializing the values at compile time for the C++. They'll end
up in the .data segment of the program, so it will be faster than
dynamic lookup, but if that's a problem we could still improve it later.

This also has an impact on the output size, for example the JS output of
the example is 154 bytes larger.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>