matrix: Implement safe versions of get()/set().
The previous implementation of get() and set() didn't check Matrix
bounds, because it didn't have a proper default value to resize the
Matrix on the fly so it can match the bounds.

So, get() now returns the default value if it is out of bounds and set()
resizes the Matrix accordingly.

Of course, the internal Matrix methods use the unsafe_ versions for
speed and in case of zip() we do our own bounds checking there.

Also add a small test case to check whether extraction of things which
are completely out of bounds works properly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>