neko terminal: Get terminal size using CPR CSI.
The CPR escape sequence is ECMA-48 compliant and in VT100, so I think
it's a good fallback and even a better one than using sys.io.Process()
to call stty or tput.

Of course, if we don't have a compliant terminal, the application will
just wait until the user presses enter, because we don't have something
like isatty() available in Neko.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>