Factor out low-level term handling into RawTerm.
Now, all that untyped crap should be properly hidden away, although
we're not quite finished with everything yet.

As for the implementation, we new set raw keyboard mode with cfmakeraw()
and re-enabling the ISIG flag, so we get a SIGINT whenever someone
presses CTRL + c.

This is very useful at least when the application has no key handler.
Also having the terminal in raw mode helps a lot in implementing the
keyboard handler.

Unfortunately, all that isn't yet available for the other targets.
Although we can implement it for JavaScript (only because the drawing is
graphically), I think we won't get around writing an NDLL for Neko :-/

Signed-off-by: aszlig <aszlig@redmoonstudios.org>