Add preliminary key handler for C++ target.
Processing keypresses is done on every single frame, so the current
implementation isn't very efficient. In the end we might want to have an
event loop based on select() and branch out to the key handler *only* if
there were real actions.

This now needs a new Application method called on_keypress(), which I've
already provided in the example application.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>