Motion: Add a function to follow *into* Objects.
Yes, you heard correctly, it's *into*, because we don't have a
distance_to() function that checks boundaries yet.

Here I fell into a mischievous death trap wondering why my equation for
deceleration didn't work but found out, that I should have switched the
whole library to be based on float (vectors) long ago.

I was constantly wondering why the follower was surpassing the target,
because I forgot to add the delta float value to the *integer*
positions, duh!

Apart from that, the implementation is neither very performant, nor is
it very shiny and nice. It would be a good idea for later[TM] to arrange
the forces list into a tree-like Enum, so we can use them like
functional combinators.

But right now, I have to meet a deadline, so this has to wait.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>