Font: Resize internal font from 8x16 to 9x16.
The reason for this is that the terminal that we want to emulate really
has blocks of size 9x16. So far we just used Font.WIDTH + 1 to work
around this issue, but after testing with background colors, we either
would have to do more workarounds or really resize the font.

With this commit, it's done and the workarounds are removed as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>