tests: Calculate width/height on assert_area.
It really doesn't make sense to specify width and height every time we
want to write an assert on a particular area, if we implicitly provide
the area by the array of strings we're passing as the expected value.

Also, if one of these values should be incorrect, weird wrapping of rows
can happen, which in turn could turn some tests useless.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>