tests: Bail out on irregular rows in assert_area.
We definitely don't want to have this scenario, because as mentioned in
the previous commit, odd things _will_ happen if we're trying to match
against matrices with odd row lengths.

So now the test is failing if the assertion method is improperly used,
because we don't want a weak policy on tests and rather accept slower
test runs than being permissive.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>