AnimationParser: Make parse_file() macro-only.
Right now this function is really only useful for macro evaluation time,
because that's where we pick up our animations and pull them into an
ASCII matrix.

Of course, this could be made available to the C++ and Neko targets as
well, but the functions that are called from within parse_file() are
only available during macro evaluation anyway, so we can lower this
constraint later if we choose to make it available at runtime.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>