utils: Add a Misc class with a binomial function.
I've used this originally to generate accurate test data for Bezier
curves, but it might come in handy some day, so I'm adding it to a new
Misc class which I'm going to populate with all sorts of (inline) helper
functions that _might_ come in handy.

Right now, this only contains the binomial function.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>