geom: Add a new preliminary Path abstract type.
My final goal is to add support for Bezier curves, but in order to do so
with a certain precision, we need to subdivide the curve and draw paths
on any "flat" edges.

So in order to achieve this goal, we need to have a way to do reliable
path drawing, which this implementation does. Although it could be
optimized quite a bit, the implementation is a (not very optimized)
derivative of the algorithm described in:

https://en.wikipedia.org/wiki/Bresenham%27s_line_algorithm

However, I'm not quite fond of the use of classes to do the iteration,
but using that it might come in handy if we want to use the same
algorithm to draw circles.

Right now, the only method is map2matrix, which calls a function for
every matrix cell along a line but *without* duplicate calls.

The tests also ensure that we won't hit duplicate calls, because every
character is a letter/number from a sequence.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>