geom.Path: Turn Bresenham class into a function.
We really don't need a class for the Bresenham's line algorithm and thus
also don't need to abuse end values by setting them nullable in order to
prevent using of a boolean flag or other weird workarounds.

In the end we don't even expose the algorithm to users of the API,
because if we want to draw a line we can still use a Path with two
nodes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>