geom: Implement basic Bezier class.
Currently it's only an implementation of the De Casteljau's algorithm
and a series of fixed steps, where we derive a path from it. Judging
from that the algorithm can't be very efficient but we're going to
change that very soon as soon as we find a nice approach to subdivide
the parametric function.

Also, the assert_bezier() method from the test cases pretty much matches
the implementation of assert_path() in PathTest.hx.

As for our more efficient implementation, let's try to add a few more
test cases with the current implementation in order to make sure the
more efficient implementation will work correctly and is as precise as
the current one.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>