geom.Bezier: Merge into geom.Path.
We are deriving a Path type out of a Bezier curve anyway, so it really
doesn't make sense to add a separate class for Bezier curves. Of course
we could add other attributes for manipulating curves as well and
implement some caching, but we can still do that as soon as we need it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>