geom.Path: Add methods to append PVectors.
Those are more or less the same as Array.push() but they're called
add*() like in the List type because we want to disambiguate it from the
underlying type (the type might switch to List someday anyway).

Also, the return value is that of the current Path, so we can chain them
like "new Path().add(10, 20).add(20, 30)".

Signed-off-by: aszlig <aszlig@redmoonstudios.org>