example: Follow *into* the car instead of beneath.
Regression introduced by eeb1a7fc45fc8dc51790bbded9907098092292b0.

The regression was caused by the Motion object defaulting to follow
beneath a certain object, which causes the example to stall and not let
the dragon move at all.

Of course, even if following *beneath* the car, the dragon shouldn't
stall, which is another bug we're going to fix later.

But for now, let's get back to the initial behaviour where the dragon
sprite follows *into* the car.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>