geom.Matrix: Rename into geom.Raster.
Wanted to do this for a long time, the first reason being that it may be
confused with mathematical matrices and second, we really use this only
for rasterization.

It also contains only methods for rasterization rather than providing
methods to do matrix addition/multiplication/whatnot in the mathematical
sense.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>