default.nix: Use DCE for compiling tests.
Although this only gains very little speedup, it's at least a few
seconds and that's better than nothing.

Shady and unscientific benchmark was done using:

time nix-build -E '
    with import <nixpkgs/lib>;
    overrideDerivation (import ./.) (const { doCheck = false; })
'

And the results without DCE were:

0.50s user 0.15s system 0% cpu 1:33.33 total

With DCE:

0.46s user 0.17s system 0% cpu 1:28.32 total

So roughly 5 seconds, but as said, this test isn't very precise because
we would need benchmark within the builder.

Apart from that, I initially wanted to use DCE for building tests but
simply forgot to do it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>