geom.Path: Determine number of steps for curves.
Since we have cubic curves now, we can easily calculate the number of
required steps by deriving a quadratic curve of the cubic one.

This has a huge impact on curve drawing performance, compare the runtime
for the tests with the static step size:

./test.hxml  8.82s user 0.04s system 99% cpu 8.873 total

With our new method:

./test.hxml  2.07s user 0.03s system 99% cpu 2.099 total

Of course, we can still improve this even further, because quite a lot
of these steps aren't really necessary as we're using paths to draw
straight lines. Also, for cusps it might be a better idea to actually
use elliptic/parabolic approximation.

But for our current use case I think it's enough and we can still
improve it later when we need it.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>