term.Canvas: Implement JavaScript key handler.
It's a quite dirty implementation which closely resembles the C++
variant but can only return Char types and doesn't handle anything like
arrow keys at all.

Also, we don't have a way to distinguish between key up and key down,
which we usually have in JS but because we don't have a way to do it
within the curses implementation we only handle keypress events at the
moment to simulate the curses behaviour.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>