FrameAreaParser: Fix parsing of vertical boxes.
If we have boxes directly on top of each other, both were merged into
one big box. This didn't allow us to do things like this:

: foo :
| o o |
| `-' |
: bar :
| o o |
| .-. |

So for animation frames this would mean that we need to encode colors
either side-by-side or like this:

: plain :
| ..... |
 :  red  :
 | ..... |
  : green :
  | ..... |
   : blue  :
   | ..... |

Which obviously is very ugly and is fixed by not merging borders if the
current border already has a heading, one or more content characters and
the border which we're going to merge has a header.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>