Animation: Allow animations to be stopped.
This can be more or less worked around by setting fps to a very small
value like 0.0000000001, but that will still consume more CPU cycles
than a real stop() method which we now have.

In order to resume the animation again, we simply need to set a new fps
value.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>