AnimationParser: Match refchar on ordinal only
Whenever we have color information in conjunction with reference
characters, the symbol representing the reference character doesn't
match with the actual symbol inside our frame area.

We now compare the symbol with the reference character only with the
ordinal property and replace it with a null character, keeping color
information.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>