term/Interface: Refactor draw_char into draw_area
So far we have passed single characters to the frontend, which worked
fine for very small animations but falls short if we need to update a
large set of sprites.

The refactor currently doesn't change any of the functionality, but
moves the underlying character handling one level down to the actual
terminal implementation.

Our goal here is that for example for canvas rendering we can cache full
sprites into an offscreen buffer and thus save a bunch of blitting
operations, which are - after profiling - the main bottleneck in our
rendering engine.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>