raster: Improve zip() method
The zip() method so far has used the map() method to build up the new
Raster. However, this is inefficient because this loops based on width
and height even though we already made sure that both Raster objects
have the same size.

We now only use a flat loop of just the data attributes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>