Remove a bunch of unnecessarily inlined methods
Let's do it the other way around and only inline methods whenever it's
really needed, because inlining to much will grow the generated code and
also affects performance in some cases.

So in terms of code size we're now down to roughly half of the code that
was generated prior to this commit.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>