object: Don't z_reorder() if z value is unchanged
Reordering the z value is quite expensive, because it means sorting of
all of the sprites in the tree.

So let's only reorder if the z value was really changed and also don't
set all children dirty, because if the current rectangle gets redrawn,
it won't cause any additional overlap than the size of the current
sprite anyway.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>