Factor out rendering process into own interface
We're going to introduce renderers more suitable for pixel-based
displays, such as our JavaScript canvas renderer.

Right now, rendering is solely character-based and in order to draw
whole sprites in one BLIT we need information about individual sprites
available in the renderer, so it can rasterize all the characters into a
cache and BLIT the whole cached pixel buffer onto the display.

This is laying the groundwork for the real implementation.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>