geom: Add a new class for bit maps
This builds upon Raster and is basically an array of ints where each bit
can be set and unset individually.

Right now the implementation is pretty basic, except for a method to set
bits based on a Rect, which is there because we might use it to merge
areas of dirty rects in the Surface.

I'm not quite sure whether the latter would be the right approach, but
even if we don't end up using it for this, it can still be handy in
creating masks.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>