surface: Merge all dirty rects and draw at once
So far every update of a single Sprite has triggered a redraw, so if you
have a lot of intersecting and/or subsprites the whole chain of redraws
gets quite expensive.

I've compared the performance with a small particle test and where I had
around 4 FPS before (using the JS target), we're now at around 40 FPS,
which is a very significant improvement.

Unfortunately despite this improvement, we still need to ramp up
performance by a whole lot in the JS target.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>