rect: Add a few impure methods
While I pretty much prefer pure functions, there are some occasions
where we really can't use them. One of them is if there is a GC
involved and we need to create as few garbage as possible.

So aside from union_() and intersection_(), we now also allow to
directly set x, y, width and height directly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>