Add a generic Box sprite generator
When creating layouts we usually need a lot of boxes, so having a
dedicated class which deals with creation of rectangles with borders is
very handy here.

Of course, if it comes to layouts a lot of other things need to be taken
care of, like for example padding and margin. But that's a thing for
future developments.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>