utils: Add a new Preloader class and tests
This is much more involved than I'd like it to be, because it seems that
we can't simply create code within macros and also reference it from
different locations.

So what we're doing now is that for every Preloader.preload() call, we
create a class called PreloaderSteps_N where N is a number that's
incremented on every call. The class that's created is just a
modification of the base PreloaderSteps class with the values added and
a destruction of the previous version of PreloaderSteps.

We also @:native(PreloaderSteps) here, so that the compiled output will
just boil down to PreloaderSteps instead of the last PreloaderSteps_N
version.

Unfortunately, I haven't found a good way to directly reference the
final version of PreloaderSteps using macros, so I'm using reflection to
resolve the class name at runtime.

While this is a bit ugly, it shouldn't bother us too much, because the
preloader is for things that are too slow to be processed while
rendering, so the overhead of reflection won't hurt us too much there.

As previously mentioned, we now use the build_single() macro from our
test Runner class and also change the default.nix to run all of the
preloader tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>